# Conflicts:
#	src/main/java/commands/VierGewinnt.java

+++ src/main/java/commands/Stop.java
@@ -6,6 +6,7 @@ import java.util.TimerTask;
 
 import command.CommandHandler;
 import command.CommandManager.ParsedCommandString;
+import core.PermissionCore;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.entities.Message;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
@@ -13,11 +14,14 @@ import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 public class Stop extends CommandHandler {
 
 	public Stop() {
-		super("stop");
+		super("stop","stop","stops the bot");
 	}
 
 	@Override
 	public void execute(ParsedCommandString parsedCommand, MessageReceivedEvent event) {
+
+		if (PermissionCore.check(4,event))return;
+
 		event.getMessage().delete().queue();
 
 		Message msg = event.getTextChannel()
